Updated the type of the FileInfoUtils_InsertFileInfoIntoArray function

+++ src/diagnostics_component/utils/file_info_utils/inc/file_info_utils.h
@@ -35,7 +35,7 @@ bool FileInfoUtils_InsertFileInfoIntoArray(
     FileInfo* sortedLogFiles,
     size_t sortedLogFileLength,
     const char* candidateFileName,
-    unsigned long sizeOfCandidateFile,
+    long long sizeOfCandidateFile,
     time_t candidateLastWrite);
 
 EXTERN_C_END
