+++ src/diagnostics_component/utils/file_info_utils/src/file_info_utils.c
@@ -45,7 +45,7 @@ bool FileInfoUtils_InsertFileInfoIntoArray(
     FileInfo* sortedLogFiles,
     size_t sortedLogFileLength,
     const char* candidateFileName,
-    unsigned long sizeOfCandidateFile,
+    long long sizeOfCandidateFile,
     time_t candidateLastWrite)
 {
     if (sortedLogFiles == NULL || sortedLogFileLength < 1 || candidateFileName == NULL || sizeOfCandidateFile == 0)
